Fix race bug and add race test (#35)
* fix race bug

* Fix CircleCI and address a test race condition (#36)

* Build within the Go path.

During CI runs when importing goftp/server from within tests, the
version in $GOPATH that `go get` obtains is used instead of the
locally checked out version. This sets a symlink in the $GOPATH to the
checked out version to make sure tests are operating on the proper
code.

* Fix race condition in test

The test would sometimes get connection refused errors as the server
takes a little time to start listening. This includes retry logic to
allow errors for a half second before failing the test.