Fix "bind: address already in use" message when using passive mode (#50)
Before this change, conn.PassivePort() could choose a port that was
already in use if the `PassivePorts` option is in use.  This is
increasingly likely on busy servers.  This would lead to messages in
the log like this:

2018/09/08 12:29:06 listen tcp :31351: bind: address already in use

This change detects this error and tries a new port number (up to 10
times) to try to find a port which isn't in use.